Avoiding UIViewExtensions code duplication, fix for Android ContextHelper potentially null
