+++ src/Uno.UI/Extensions/UIViewExtensions.iOSmacOS.cs
@@ -287,26 +287,13 @@ namespace AppKit
 			view.Frame = new CGRect(view.Frame.X, view.Frame.Y, width ?? view.Frame.Width, height ?? view.Frame.Height);
 		}
 
-		public static _View FindFirstResponder(this _View view)
-		{
-#if __IOS__
-			if (view.IsFirstResponder)
-#elif __MACOS__
-			if (view.Window.FirstResponder == view)
-#endif
-			{
-				return view;
-			}
-			foreach (_View subView in view.Subviews)
-			{
-				var firstResponder = subView.FindFirstResponder();
-				if (firstResponder != null)
-				{
-					return firstResponder;
-				}
-			}
-			return null;
-		}
+		/// <summary>
+		/// Finds first responder in view
+		/// </summary>
+		/// <param name="view">View</param>
+		/// <returns>First responder view</returns>
+		public static _View FindFirstResponder(this _View view) =>
+			Uno.Extensions.UIViewExtensions.FindFirstResponder(view);		
 
 		/// <summary>
 		/// Finds the nearest view controller for this _View.
