+++ src/Uno.UWP/Globalization/Language.iOS.cs
@@ -1,5 +1,6 @@
 ﻿#if __IOS__
 using UIKit;
+using Uno.Extensions;
 
 namespace Windows.Globalization
 {
@@ -12,28 +13,11 @@ namespace Windows.Globalization
 				UIView responder = null;
 				if (UIApplication.SharedApplication?.KeyWindow?.RootViewController?.View != null)
 				{
-					responder = FindFirstResponder(UIApplication.SharedApplication.KeyWindow.RootViewController.View);
+					responder = UIApplication.SharedApplication.KeyWindow.RootViewController.View.FindFirstResponder();
 				}
 				var inputMode = responder?.TextInputMode;
 				return inputMode?.PrimaryLanguage ?? "";
 			}
-		}
-
-		private static UIView FindFirstResponder(UIView view)
-		{
-			if (view.IsFirstResponder)
-			{
-				return view;
-			}
-			foreach (var subView in view.Subviews)
-			{
-				var firstResponder = FindFirstResponder(subView);
-				if (firstResponder != null)
-				{
-					return firstResponder;
-				}
-			}
-			return null;
 		}
 
 		public static bool TrySetInputMethodLanguageTag(string languageTag) => false;
